bug fixes
fixed the closeAllModals utility so that it actually closes the modals.
fixed shipping addresses so they're editable in my account area.
fixed address edit so if change is made and save is pushed before blur,
value is still found and saved (was saving blank previously)
fixed meta tracking to look for meta instead of META on URI
fixed customer page so loadingBG class is removed after page is done
loading.