bug fix for modal executed in showContent
showContent automatically scrolled to the top of the page, however this
caused issues if a a link was at the bottom of the page and opened a
modal. so now, you can inherently turn it off by passing
performTransition:false as part of the second param in show content.
Also, the app is smart enough to default to no transition for login if
user is not logged in. a passed param will override any default behavior
though.