new category list render format
A new category list render format that is more in line with the new
product list format (all self-contained). Allows for a faster render for
a simple list and the request to be tied to the type of data needed.

started work on updates to the add to cart functions.

Changed the rendering functions so the 'replace' no longer occurs (in
controller).

Fixed issue with new product list due to 'replace' mentioned above
(controller) so timeout workaround (which was temporary fix) is
resolved.