Fix prodlist bug. some request optimization
Updated model so that the contents of _tag for each request are not
sent. Instead, handleResponse will look it up and add it as _rtag (just
like API would if it was sent).

Fixed bug in product list where prodlist code got executed even if the
list had no value.

turned off some debugging code.