updates to checkout. 2 new extensions
New version of the API with new cart object (much improved) meant
changes to all of the templates and checkout extensions to point to the
new fields.

New extension for powerReviews
New extension for MagicTool Box (magicThumb and magicZoomPlus)

Found a few references to myControl in the examples.
Fixed a few bugs in some message handling.
Changed how giftcards and coupons are handled on the cart page. The cart
update request now only occurs if the coupon/gc add is successful