Added State and History of the World
Each time showContent is run, state of the world changes to the page
info obj. This will allow callbacks to ensure the state hasn't changed
and whether or not to render (not built in yet).

History of the world is an array of what the 15 most recent states are,
where spot 0 is always the most recent.

Added a new templateFunction state known as 'onDeparts' which gets
executed when a user leaves a page. Push onto this just like you would
onInits or onCompletes