admin - media lib
redid how folders are rendered so that they're all generated at the
outset, instead of subfolders being generated onclick. This was more
effecient (only loop through folders once) and allowed for a link to be
generated to open a specific folder several folders deep.

Closer to supporting the deletion of a folder. The delete works, but the
UI behaviors need some love.