admin - fixes and features
Added a 'select domain' modal that appears on admin init (unless one is
already set in localstorage, memory or on URI).

update to media lib so after a multiple file upload, the command to
update the folder list only gets executed once.

Fix to auth to remove 'secret' as part of MD5.

Fix to remove some duplicate messaging that was occuring.