admin - bug fixes
When a template.html file is loaded, release is appended on URI to
decrease likelyhood of caching.
Script loading already appended release, but didn't assign a key (_v=)
which it does now.
Start of utilities page as a template.
New template based setup page now runs handleURLRewrites so links work.