new renderformat and admin function
A couple new functions to make things easier for developers.
renderFormat processList allows any object to be passed in as a
data-bind var and this'll loop through and use loadsTemplate on each.
This is intended to replace stuffList and object2Template

Also added an admin function for processing a form and submitting it as
an api request. see docs for how to use.