quickstart.js, store_search.js - Exploring complex elastic queries
Created new query-building function in store_search, that is accessible
by using the searchLink renderFormat, and adding the data-bind
attribute: elasticAttr equal to a space-separated list of values to be
keyed over

This is some very rough work, intended only to explore the
possibilities.  I'm not convinced that the data-bind is the best way to
approach it, but having more complex query builders will certainly be
useful.  For example, this query filter would allow a list of checkboxes
for fields to return a more refined query over only sku, or prod_name.

Future expansions should consider adding range filters, for things like
price to allow further refinement of search results.

One area that needs to be discussed further is that of the data-pointer
and its syntax, as well as how recent searches are stored and
referenced.