admin media - bug fix/enhancement
Updated media library and how lazy load works. Instead of running the
button conversion script over the entire dom element or within the loop
over each li, a param was added to lazy load to support running a
function as each image is loaded. Much more efficient. Tested on a
folder with 1600 images and it worked fine.