quickstart - new non-id based tab script
Added a simple tab script 'anytabs' which is a fairly dumbed down tab
script but one that does NOT use id's so it's more app friendly. details
on how to use are in anytabs.js at top (in comment section) and init.js
and the product template in quickstart have already been updated to use.