quickstart - enhanced search results functionality
The search results page will now open a 'preview' pane when a result is
clicked. The template used is the same as quickview (but doesn't have to
be) and allows the user to get a more detailed view on the page without
leaving. 'next' and 'previous' buttons also exist to fascilitate easy
browsing.