quickstart enhancements and bug fix
Fixed a bug in the product template which was causing accessories to not
be tabbed.

Added a new feature to the product detail page (when arriving from a
category page) that will add previous and next buttons to the page to
allow the users to easily move forward/back through the list of product.