admin - bug fix and rename for preferences
The get and set call for preferences has been changed to dpsGet and
dpsSet (device persistent settings). The functions can be used to get
and set UI settings (such as a panel state) that will be retained in
local storage and used on the users next visit to the page/app (such as
what order filters were last used).

There was a bug in this script previously that caused 'unchecked' items
to remain in the localstorage object.