admin prodEdit - enhancement
From a search results list in the product editor, clicking 'edit' will
now open a product results tab which moves the search results to a 'tab'
hanging on the left side of the screen, which will follow the user
through the product edit experience. leaving the product tab will clear
the results tab.