new cart extension 'quickadd' for add to cart
An alternative action for cart display after adding an item to the cart.

Instead of the full-blown cart modal, this will animate a small element
in from the right and display either the item added (if successful) or
the error message.

Also will auto-retract after five seconds and includes a shop more
button (which also retracts the element) and a checkout button.