FF lockup bugfix
Firefox locks up when rendering large data from localstorage due to the
fact that it doesn't release the single thread of execution.  When an
API call is made, the callback is performed asynchronously and doesn't
block, however handleCallback does not.  Adding this small delay
releases the single thread to perform other operations and queues the
callbacks.