admin syndication - bug fix when switching between DSTs
There is a bug either in the tab script (which I looked for but could
not find) or the browser. Even though the container is emptied (i even
destroyed the tabs at one point), when the new editor appears, whichever
tab was previously selected stays selected. The code below triggers a
tab click but not the request code for the first tab.