admin/anyplugins - enhancements
anyplugins stickytab had an issue of the width on the closed tab being
too big and causing a z-index issue in parts of the UI.

batch manager stickytab will now only appear when in the batch manager
(would previously 'stick' in reports).

giftcard interface is now appificated, but not linked.