admin - enhancements and bug fixes
Fixed a bug in ebay 'refresh listings'. wrong param key being sent.
Updated the prodlist functionality in the template editor. There's now
two functions. 1 for conjuring product which saved the result set to a
hidden id. The second is for actually rendering the product list. This
potentially puts a lot more control into the hands of the wizard
developer.