admin - bug fixes and enhancements
Fixed a bug in anytable that was causing it to break if it was applied
twice to the same table.  it'll now check to see if it's already been
added and, if so, handle it accordingly.

Migrated batch job manager to use DMI.
Upgraded most of the other functions in there to use inline dispatches
instead of 'calls' and removed the batch calls from the admin extension.