ebay item specifics - nested/related fieldsets fix
Sometimes fieldsets are not nested, but still related - and we need to handle this.
See "DVDs & Movies -> VHS Tapes" for how this works
Changing 'Genre' automatically shows/hides checkboxes inside 'Sub-Genre' - but they're not nested!

Now if you see "DVDs & Movies -> DVDs & Blu-ray Discs" category
there we have the same Genre and Sub-Genre fieldsets - but this time they're nested.
Fixed jquery.ebay-specifics-form.js to work both with nested AND not nested but related.