enhancements and bug fixes
anytabs now uses delegated events for clicks.

added a new utiltity for handleCommonPlugins, which will use the
applyAnycb (or whatever the supporting class/plugin is) and apply the
plugin.

Added a utility for normalizing event types. 'blur' and 'focus' are
actually onfocus and onfocusout, which aren't as pretty or in sync with
jquery event names.

added support for blur and focus in event delegator.

Some generic blur/focus events for resizing textarea.

Fixed bug in navcats that caused items that should be pre-checked to not
be if navcatDetail not in memory/LS already.

More work on product editor. Cleaned up some tab content and also have
navcats, flexedit and social all loading their respective content now.