admin - enhancements
Webdoc is now a supported article type for trainer.
Cleaned up admin-init to not include some files that are no longer in
use.

Added classes for the message states. These classes will be used both in
anymessage and in the messages console linked to from the header.

added an icon-chart file to facilitate easy testing for icon classes.