core enhancement
A major change to the init process. This change allows for multiple
instantiations of the app to be open at once. Cleans up the app init
file substantially as well, moving everything but the preloader code
(which is app specific) from the init into the controller.
loadscript.js is no longer necessary, as the functionality is now within
the controller.