order create - bug fix
An issue w/ the blur event on some inputs, such as zip, resetting the
panel w/ the save button. If you went from 'zip', directly to save, the
click on the save registered the blur event and not a click, causing the
app to appear to have saved but not saved. Moving the buttons outside
that panel solved the problem.