admin - enhancement to product task list
Ran into an issue where if the product tab content hadn't been rendered
during the session, the product task list didn't work right. To fix,
items are stored and added to the list at render time.  The enhancement
part of this is that the product task list is now persistent between
sessions.