tlc - update to how 'globals' are handled
The 'globals' are the statement-specific hash that is updated as the
statement is processed.  Originally, if a foreach or if/block was hit,
globals was duplicated and passed in.  That is not ideal because those
types of commands need to be able to update that globals object.  They
can now.