product editor - bug fixes
In a managed variation, enabling an option didn't trigger the save
button to unlock.
Changed the 'check for inventory-able variations' code worked to
actually look through the variations object instead of the skus. That
means it'll work if a inventory-able variation has been enabled, but no
options have been configured.