product editor - start of new sku-finder feature
For product that have a lot of inventory-able options, there was
potential for browser lockup.  This feature will detect a high number of
sku's for a given product and instead of listing them, will present a
chooser (attributes tab, variations, competition and buy.com were
impacted by this).
Works.  needs styling.